requirements.txt
# ===== API.GEO.Carbone — Requirements =====
# Python 3.10+

# Django core
Django>=4.2,<5.0
django-environ>=0.11.2

# Django GIS
django-leaflet>=0.29.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.3
django-cors-headers>=4.3.0

# Database (PostGIS)
psycopg2-binary>=2.9.9

# Admin theme
django-jazzmin>=2.6.0

# Forms
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Dev tools
python-dotenv>=1.0.0

# ===== Production (Render) =====
gunicorn>=21.2.0
whitenoise[brotli]>=6.6.0
dj-database-url>=2.1.0
